Fixed bug in Glicko2

+++ Lavasurvival/src/main/java/me/eddiep/minecraft/ls/glicko/Glicko2.java
@@ -37,7 +37,7 @@ public class Glicko2 {
         return tau;
     }
 
-    public double getDefaultRating() {
+    public int getDefaultRating() {
         return default_rating;
     }
